pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "yamlbridge"
version = "0.1.0"
description = "PyYAML-compatible loaders for vendor-specific YAML dialects"
readme = "README.md"
requires-python = ">=3.9"
license = "MIT"
authors = [
  { name = "Shoumik Gandre" }
]
keywords = ["yaml", "cloudformation", "pyyaml", "parser", "config"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
  "pyyaml>=6.0",
]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
    "pytest-html>=4.2.0",
    "ruff>=0.15.6",
]

[tool.ruff]
line-length = 100
target-version = "py39"
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = [
  "-ra",
  "-v",
  "--junitxml=.test-reports/junit.xml",
  "--cov=yamlbridge",
  "--cov-report=term-missing",
  "--cov-report=html:.test-reports/coverage-html",
  "--cov-report=xml:.test-reports/coverage.xml",
  "--html=.test-reports/unit-tests.html"
]
junit_family = "xunit2"
junit_duration_report = "total"

[tool.mypy]
python_version = "3.9"
mypy_path = ["src"]
strict = true
pretty = true
show_error_codes = true

